move helper class code inside xattrs table
