+++ include/tables/XAttrTable.h
@@ -75,6 +75,115 @@ struct XAttrRow {
 typedef vector<XAttrRow> XAttrVec;
 typedef boost::unordered_map<string, XAttrVec> XAttrMap;
 
+class FsMutationsHelper {
+  public:
+    static string to_upsert_json(XAttrRow row){
+      stringstream out;
+      out << getDocUpdatePrefix(row.mInodeId) << endl;
+      out << upsert(row) << endl;
+      return out.str();
+    }
+
+    static string to_delete_json(Int64 inodeId, string name){
+      stringstream out;
+      out << getDocUpdatePrefix(inodeId) << endl;
+      out << removeXAttrScript(name) << endl;
+      return out.str();
+    }
+  private:
+    static string upsert(XAttrRow row) {
+      rapidjson::Document doc;
+      doc.Parse(getXAttrDoc(row, true).c_str());
+      rapidjson::Document xattr(&doc.GetAllocator());
+      if (!xattr.Parse(row.mValue.c_str()).HasParseError()) {
+        mergeDoc(row.mName, doc, xattr);
+        rapidjson::StringBuffer sbDoc;
+        rapidjson::Writer<rapidjson::StringBuffer> docWriter(sbDoc);
+        doc.Accept(docWriter);
+        return string(sbDoc.GetString());
+      } else {
+        LOG_DEBUG("XAttr is non json " << row.mName << "=" << row.mValue);
+        return getXAttrDoc(row, false);
+      }
+    }
+
+    static void mergeDoc(string attrName, rapidjson::Document& target, rapidjson::Document& source) {
+      for (rapidjson::Document::MemberIterator itr = source.MemberBegin(); itr != source.MemberEnd(); ++itr) {
+        target["doc"][XATTR_FIELD_NAME][attrName.c_str()].AddMember(itr->name,
+            itr->value, target.GetAllocator());
+      }
+    }
+
+    static string removeXAttrScript(string xattrname){
+
+      rapidjson::StringBuffer sbOp;
+      rapidjson::Writer<rapidjson::StringBuffer> opWriter(sbOp);
+
+      opWriter.StartObject();
+
+      opWriter.String("script");
+
+      stringstream rmout;
+      rmout << "ctx._source." << XATTR_FIELD_NAME << ".remove(\"" << xattrname <<  "\")";
+      opWriter.String(rmout.str().c_str());
+
+      opWriter.EndObject();
+
+      return string(sbOp.GetString());
+
+    }
+
+    static string getXAttrDoc(XAttrRow row, bool isJSONVal){
+
+      rapidjson::StringBuffer sbOp;
+      rapidjson::Writer<rapidjson::StringBuffer> opWriter(sbOp);
+
+      opWriter.StartObject();
+
+      opWriter.String("doc");
+      opWriter.StartObject();
+
+      opWriter.String(XATTR_FIELD_NAME);
+      opWriter.StartObject();
+
+      opWriter.String(row.mName.c_str());
+      if(isJSONVal) {
+        opWriter.StartObject();
+        opWriter.EndObject();
+      }else{
+        opWriter.String(row.mValue.c_str());
+      }
+      opWriter.EndObject();
+
+
+      opWriter.EndObject();
+      opWriter.String("doc_as_upsert");
+      opWriter.Bool(true);
+
+      opWriter.EndObject();
+
+      return string(sbOp.GetString());
+    }
+
+    static string getDocUpdatePrefix(Int64 inodeId){
+      rapidjson::StringBuffer sbOp;
+      rapidjson::Writer<rapidjson::StringBuffer> opWriter(sbOp);
+
+      opWriter.StartObject();
+
+      opWriter.String("update");
+      opWriter.StartObject();
+
+      opWriter.String("_id");
+      opWriter.Int64(inodeId);
+
+      opWriter.EndObject();
+      opWriter.EndObject();
+
+      return string(sbOp.GetString());
+    }
+};
+
 class XAttrTable : public DBTable<XAttrRow> {
 
 public:
@@ -164,6 +273,5 @@ public:
     args[0] = inodeId;
     return doRead(connection, PRIMARY_INDEX, args);
   }
-
 };
 #endif //EPIPE_XATTRTABLE_H
